修复利息

+++ src/interest.h
@@ -1,4 +1,4 @@
 #include"banklist.h"
 
 
-double interest(struct Account p,struct Date day1);//输入用户p，存款到期日day1,返回获得的利息（两位小数）
+long long interest(struct Account p,struct Date day1);//输入用户p，存款到期日day1,返回获得的利息（两位小数）
